Working Auth

+++ api/src/main/java/io/trenchcrusade/api/warband/WarbandRepository.java
@@ -2,4 +2,4 @@ package io.trenchcrusade.api.warband;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface WarbandRepository extends CrudRepository<Warband, Integer> {}
+public interface WarbandRepository extends CrudRepository<Warband, Long> {}
